Push login & QR code

+++ esprit1718b1businessbuilder/esprit1718b1businessbuilder-client/src/main/java/tn/esprit/b1/esprit1718b1businessbuilder/app/client/main/TenderTest.java
@@ -68,7 +68,7 @@ public class TenderTest {
 		//tender.setCompanyTender(loggedUser);
 		//tender.setCategory(HelpDesk);
 		//proxyTender.save(tender);
-		long id = 3;
+		int id = 3;
 		System.out.println(proxyCompany.find(id));
 	}
